编写限购信息更新接口

+++ ec-limitation-server/src/main/java/com.weimob.saas.ec.limitation/entity/GoodsLimitInfoEntity.java
@@ -9,7 +9,7 @@ public class GoodsLimitInfoEntity {
 
     private Long limitId;
 
-    private Byte limitType;
+    private Integer limitType;
 
     private Long goodsId;
 
@@ -49,11 +49,11 @@ public class GoodsLimitInfoEntity {
         this.limitId = limitId;
     }
 
-    public Byte getLimitType() {
+    public Integer getLimitType() {
         return limitType;
     }
 
-    public void setLimitType(Byte limitType) {
+    public void setLimitType(Integer limitType) {
         this.limitType = limitType;
     }
